build: stabilisation et finalisation du projet elcollecte
- Architecture microservices complète et fonctionnelle
- Intégration de tous les services (collecte, rapport, utilisateur, projet, formulaire, média, validation, analytique, audit)
- Passerelle API (api-gateway) configurée
- Service de découverte (Eureka) opérationnel
- Messagerie asynchrone Kafka intégrée
- Base de données PostgreSQL avec migrations Flyway
- Sécurité Spring Security activée
- Documentation OpenAPI/Swagger complétée
- Interface frontend intégrée
- Docker et Docker Compose configurés
- Tests unitaires et d'intégration en place
- All services tested and working correctly

+++ elcollecte/service-analytique/src/main/resources/db/migration/V1__init_analytique.sql
@@ -4,7 +4,7 @@ CREATE TABLE stats_projets (
     total_collectes      INT        NOT NULL DEFAULT 0,
     collectes_validees   INT        NOT NULL DEFAULT 0,
     collectes_rejetees   INT        NOT NULL DEFAULT 0,
-    collectes_en_attente INT        NOT NULL DEFAULT 0,
+    collectes_en_attente BIGINT        NOT NULL DEFAULT 0,
     nb_enqueteurs_actifs INT        NOT NULL DEFAULT 0,
     taux_validation      NUMERIC(5,2) DEFAULT 0,
     updated_at           TIMESTAMP  NOT NULL DEFAULT NOW()
